メイン画面をページにした
